backend/pyproject.toml
[project]
name = "autoclip"
version = "2.0.0"
description = "Conversational AI video clipping with multimodal LangGraph pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.6.0",
    "httpx>=0.25.0",
    "google-genai>=1.0.0",
    "assemblyai>=0.20.0",
    "scenedetect[opencv]>=0.6.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autoclip"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]
